[#78] added delete property to indicate 'soft-deletion' of a product if there is a reference to invoices

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/ProductServiceImpl.java
@@ -1,12 +1,15 @@
 package at.ac.tuwien.sepm.groupphase.backend.service.impl;
 
 import at.ac.tuwien.sepm.groupphase.backend.entity.Category;
+import at.ac.tuwien.sepm.groupphase.backend.entity.InvoiceItem;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Product;
 import at.ac.tuwien.sepm.groupphase.backend.entity.TaxRate;
 import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundException;
 import at.ac.tuwien.sepm.groupphase.backend.repository.CategoryRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.ProductRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.TaxRateRepository;
+import at.ac.tuwien.sepm.groupphase.backend.service.InvoiceItemService;
+import at.ac.tuwien.sepm.groupphase.backend.service.InvoiceService;
 import at.ac.tuwien.sepm.groupphase.backend.service.ProductService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,6 +26,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.lang.invoke.MethodHandles;
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -31,12 +35,18 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
     private final CategoryRepository categoryRepository;
     private final TaxRateRepository taxRateRepository;
+    private final InvoiceItemService invoiceItemService;
 
     @Autowired
-    public ProductServiceImpl(ProductRepository productRepository, CategoryRepository categoryRepository, TaxRateRepository taxRateRepository) {
+    public ProductServiceImpl(
+        ProductRepository productRepository,
+        CategoryRepository categoryRepository,
+        TaxRateRepository taxRateRepository,
+        InvoiceItemService invoiceItemService) {
         this.productRepository = productRepository;
         this.categoryRepository = categoryRepository;
         this.taxRateRepository = taxRateRepository;
+        this.invoiceItemService = invoiceItemService;
     }
 
     @Caching(evict = {
@@ -172,9 +182,20 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void deleteProductById(Long productId) {
         LOGGER.trace("deleteProductById{}", productId);
-        productRepository.findById(productId)
+        boolean softDelete = false;
+        Product productToDelete = productRepository.findById(productId)
             .orElseThrow(() -> new NotFoundException(String.format("Could not find product with id: %s", productId)));
-        productRepository.deleteById(productId);
+        softDelete = this.invoiceItemService.findAllInvoicesItems().stream()
+            .map(InvoiceItem::getProduct)
+            .anyMatch(product -> product.getId().equals(productId));
+
+        if (!softDelete) {
+            productRepository.deleteById(productId);
+        } else {
+            productToDelete.setDeleted(true);
+            productRepository.save(productToDelete);
+        }
+
     }
 
     @Override
